requirements.txt
python-decouple>=3.8
click>=8.0.0
typer>=0.9.0
pydantic>=1.10.0 